+++ src/stewards/quality/__init__.py
@@ -1,0 +1,1 @@
+"""The MeshOps Quality Steward (hello-quality) package."""